Spanner test hygiene: move fixtures out of distro.
Also used our GrpcRendezvous alias for _Rendezvous and making
some "conditionally defined" functions into a module-level function
and a test case staticmethod.
