+++ spanner/tests/_fixtures.py
